ui_enhancements

+++ lib/event_list/controller/event_list_controller.dart
@@ -1,8 +1,10 @@
+import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:where_hearts_meet/create_event/model/event_response_model.dart';
 
 import '../../utils/consts/screen_const.dart';
 import '../../utils/controller/base_controller.dart';
+import '../../utils/dialogs/confirmation_dialog.dart';
 import '../../utils/dialogs/pop_up_dialogs.dart';
 import '../service/event_list_service.dart';
 
@@ -49,12 +51,20 @@ class EventListController extends BaseController {
   }
 
   void deleteEvent({required String eventId}) async {
-    showLoaderDialog(context: Get.context!);
-    final res = await _eventListService.deleteEventApi(eventId: eventId);
-    cancelDialog();
-    if (res != null) {
-      eventsList?.removeWhere((element) => element.eventid == eventId);
-    }
-    update();
+    showConfirmationBottomSheet(
+        context: Get.context!,
+        onTapYes: () async {
+          Get.back();
+          showLoaderDialog(context: Get.context!);
+          final res = await _eventListService.deleteEventApi(eventId: eventId);
+          cancelDialog();
+          if (res != null) {
+            eventsList?.removeWhere((element) => element.eventid == eventId);
+          }
+          update();
+        },
+        text: 'Delete Event?',
+        description: 'Are you sure to delete this event?',
+        iconData: Icons.delete);
   }
 }
